requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
